+++ Releases.md
@@ -3,6 +3,7 @@
 ## 1.12.0
 
 * updated dependency: cubx-grunt-webpackage-scaffold to version v1.8.0 - PLAT-380: "+webpackage-createElementary" and "+webpackage-createCompound" generate also demo-app for the component
+* changed "cubixx" to "cubbles"
 
 ## 1.11.2